# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CtrlTrade SaaS Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - in: query
          name: status
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/JobSummary" }
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /v1/jobs/{jobId}:
    parameters:
      - in: path
        name: jobId
        required: true
        schema: { type: string }
    get:
      operationId: getJob
      tags: [jobs]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
    patch:
      operationId: updateJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /v1/jobs/{jobId}/assign:
    parameters:
      - in: path
        name: jobId
        required: true
        schema: { type: string }
    post:
      operationId: assignJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobAssignmentInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
...
  /v1/schedule:
    get:
      operationId: getSchedule
      tags: [schedule]
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScheduleEntry" }
...
components:
  schemas:
...
    JobSummary:
      type: object
      required: [id, number, title, status, customerId, customerName, valuePence, createdAt]
      properties:
        id: { type: string }
        number: { type: string }
        title: { type: string }
        status: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        scheduledStart: { type: ["string", "null"], format: date-time }
        scheduledEnd: { type: ["string", "null"], format: date-time }
        assignedUserId: { type: ["string", "null"] }
        assignedUserName: { type: ["string", "null"] }
        valuePence: { type: integer }
        createdAt: { type: string, format: date-time }
    Job:
      type: object
      required: [id, number, title, status, customerId, customerName, valuePence, createdAt]
      properties:
        id: { type: string }
        number: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        quoteId: { type: ["string", "null"] }
        scheduledStart: { type: ["string", "null"], format: date-time }
        scheduledEnd: { type: ["string", "null"], format: date-time }
        addressLine1: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        postcode: { type: ["string", "null"] }
        assignedUserId: { type: ["string", "null"] }
        assignedUserName: { type: ["string", "null"] }
        assignedVehicleId: { type: ["string", "null"] }
        valuePence: { type: integer }
        createdAt: { type: string, format: date-time }
    JobInput:
      type: object
      required: [customerId, title]
      properties:
        customerId: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string }
        scheduledStart: { type: string, format: date-time }
        scheduledEnd: { type: string, format: date-time }
        addressLine1: { type: string }
        city: { type: string }
        postcode: { type: string }
        assignedUserId: { type: string }
        assignedVehicleId: { type: string }
        valuePence: { type: integer, minimum: 0 }
    JobAssignmentInput:
      type: object
      properties:
        assignedUserId: { type: ["string", "null"] }
        assignedVehicleId: { type: ["string", "null"] }
        scheduledStart: { type: ["string", "null"], format: date-time }
        scheduledEnd: { type: ["string", "null"], format: date-time }
    ScheduleEntry:
      type: object
      required: [jobId, title, status, start, customerName]
      properties:
        jobId: { type: string }
        title: { type: string }
        status: { type: string }
        start: { type: string, format: date-time }
        end: { type: ["string", "null"], format: date-time }
        assignedUserId: { type: ["string", "null"] }
        assignedUserName: { type: ["string", "null"] }
        customerName: { type: string }
...